+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add badge to spells in list showing damaging spells
+
 ### Changed
 
 - All damage lowered by 2 dice levels. Bashing is d4, aggravated is d6, lethal is d8